% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategus.R
\name{create_options}
\alias{create_options}
\title{create options for running the Strategus flow in D2E}
\usage{
create_options(
  study_id = "",
  upload_results = FALSE,
  update_results_schema = TRUE,
  study_name = ""
)
}
\arguments{
\item{study_id}{string value indicating the study ID to be used in the flow run (maintained for backward compatibility, but study_name should be used going forward)}

\item{upload_results}{boolean value indicating whether to upload results after the flow run
(default is FALSE)}

\item{update_results_schema}{boolean value indicating whether to update results schema
before uploading results (default is TRUE). Any existing results in the schema will be
dropped before uploading new results.}


\item{study_name}{string value indicating the study name to be used in the flow run
(default is an empty string). If provided, this will override the study_id parameter.}
}
\value{
Response object with options for the flow run
}
\description{
This function creates options for a flow run in Prefect in D2E.
So far, the only option is to upload_results.
}
