% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortDefinitionSet.R
\name{get_cohort_definition_set}
\alias{get_cohort_definition_set}
\title{Get the cohort defintion set by a list of cohort Ids}
\usage{
get_cohort_definition_set(cohortIds, generateStats = FALSE)
}
\arguments{
\item{cohortIds}{a list of cohortIds}

\item{generateStats}{a booloean to generate statistics}
}
\value{
CohortDefinitionSet object
}
\description{
This function fetches cohort defintions and converts them
to cohort definition set object, compatible to use in
OHDSI Strategus and Cohort Generator R packages. It also uses
Circe package for converting the cohort JSON to Sql
}
