% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategus.R
\name{run_strategus_flow}
\alias{run_strategus_flow}
\title{Start running the Strategus flow in D2E}
\usage{
run_strategus_flow(dataset_code, analysisSpecification, executionSettings)
}
\arguments{
\item{dataset_code}{is a string, which is the unique identifier of dataset in D2E}

\item{analysisSpecification}{AnalysisSpecification object created by Strategus::createEmptyAnalysisSpecificiations}

\item{executionSettings}{ExecutionSettings object created by Strategus::createCdmExecutionSettings}
}
\value{
Response object with flow run details, id and status
}
\description{
This function creates a flow run in Prefect in D2E.
Internally the flow run starts a network study analysis on the given dataset
using the OHDSI Strategus R package.
}
