% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortDefinitionSet.R
\name{create_cohort_definition}
\alias{create_cohort_definition}
\title{Create the cohort defintion}
\usage{
create_cohort_definition(name, description, cohort_definition_json)
}
\arguments{
\item{name}{a string value indicating the name of the cohort definition}

\item{description}{a string value indicating the description of the cohort definition}

\item{cohort_definition_json}{a list object containing the cohort definition JSON.}
}
\value{
CohortDefinition object
}
\description{
This function creates cohort defintion compatible to use in R packages.
It accepts name, description and a cohort definition JSON object
with or without an expression field.
}
