% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategus.R
\name{create_options}
\alias{create_options}
\title{create options for running the Strategus flow in D2E}
\usage{
create_options(study_id = "", upload_results = FALSE)
}
\arguments{
\item{study_id}{string value indicating the study ID to be used in the flow run}

\item{upload_results}{boolean value indicating whether to upload results after the flow run
(default is FALSE)}
}
\value{
Response object with options for the flow run
}
\description{
This function creates options for a flow run in Prefect in D2E.
So far, the only option is to upload_results.
}
